bugfix: fixed the TEST_NGINX_BINARY environment for specifying nginx binary path.
Prior to this patch, specifying $TEST_NGINX_BINARY has the
effect of iterating over $PATH segments, and concatenating the specified
path. Ex:

    $ TEST_NGINX_BINARY=/some/path/nginx prove t/
        testing: /usr/loca/bin//some/path/nginx
        testing: /usr/bin//some/path/nginx
        etc...

Hence this variable only has effect if it specifies a command name
available in the $PATH, which is contrary to that part of the
documentation if I interpret it correctly:

> Can be used as an alternative to setting PATH to run a specific
> nginx instance.

This behavior I believe was broken by eea73704.

With this patch, we ensure the given $cmd is an executable file.

    $ TEST_NGINX_BINARY=/some/path/nginx prove t/
        testing: /some/path/nginx

Signed-off-by: Yichun Zhang (agentzh) <agentzh@gmail.com>