bugfix: re-enable usage of TEST_NGINX_BINARY
Prior to this patch, specifying $TEST_NGINX_BINARY has the
effect of iterating over $PATH segments, and concatenating the specified
path. Ex:

    $ TEST_NGINX_BINARY=/some/path/nginx prove t/
        testing: /usr/loca/bin//some/path/nginx
        testing: /usr/bin//some/path/nginx
        etc...

Hence this variable only has effect if it specifies a command name
available in the $PATH, which is contrary to that part of the
documentation if I interpret it correctly:

> Can be used as an alternative to setting PATH to run a specific
> nginx instance.

This behavior I believe was broken by
eea73704b195292a6acad8a75f4b069a02fa86c0.

With this patch, we ensure the given $cmd is an executable file.

    $ TEST_NGINX_BINARY=/some/path/nginx prove t/
        testing: /some/path/nginx