bugfix: ensure servroot is cleaned up when starting NGINX for the first time.
Since 56733f3, we do not remove all files in the servroot/html
directory, which is necessary for existing test cases when running in
HUP mode.

However, an existing issue with NGINX (trac #753
https://trac.nginx.org/nginx/ticket/753) causes the SIGQUIT signal
(graceful exit) to not remove listening unix socket files. This causes
subsequent tests run to fail due to NGINX to being able to bind since
the file already exists.

This fix ensures that we remove all files in the servroot/html
directory before running a test suite.