Util.pm: preset *temp_path directives to subdirectories of servroot
This avoids errors when NGINX, launched as non-root, tries writing
various temporary files into default compiled-in directories that are
only writable by root.

This commit sets directives that control where the temp files are saved,
to point to subdirectories of the temporary server root, which is
writable by the user executing NGINX.

Since some test for modules like lua-nginx-module rely on specfic
hardcoded line numbers, the added directives are placed on the same line
as previously-present 'access_log'. This way, line numbers do not
change.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>