Util.pm: specify error log on command line
When no error log is specified on the command line, NGINX logs errors to
the compiled-in error log (irrespective of whether the error log is
specified in the configuration file), which is most often writable only
by root. When tests are run as unprivileged user, NGINX can not write
the log file and, therefore, fails to start.

This commit makes Test::Nginx set the error log to
$ServRoot/logs/error.log when the configuration file is specified, thus
overriding the default path.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>